modified the controllers test

+++ src/main/java/ma/emsi/blog/controller/CommentaireController.java
@@ -26,35 +26,47 @@ public class CommentaireController {
 
 	@PostMapping(value = "/commentaire", produces = { "application/json", "application/xml" }, consumes = {
 			"application/json", "application/xml" })
-	public ResponseEntity<Void> add(@RequestBody Commentaire p) {
-		commentaireService.create(p);
+	public ResponseEntity<Void> add(@RequestBody Commentaire commentaire) {
+		commentaireService.create(commentaire);
 		return ResponseEntity.noContent().build();
 	}
 
-	@GetMapping(value = "/commentaires", produces = { "application/json", "application/xml" }, consumes = {
-			"application/json", "application/xml" })
+	@GetMapping(value = "/commentaires", produces = { "application/json", "application/xml" })
 	public ResponseEntity<List<Commentaire>> findAll() {
 		List<Commentaire> result = commentaireService.findAll();
 		return ResponseEntity.ok().body(result);
 	}
 
-	@GetMapping(value = "/commentaire/id/{id}", produces = { "application/json", "application/xml" }, consumes = {
-			"application/json", "application/xml" })
+	@GetMapping(value = "/commentaire/id/{id}", produces = { "application/json", "application/xml" })
 	public ResponseEntity<Commentaire> findcommentaire(@PathVariable int id) {
 		Commentaire result = commentaireService.findById(id);
-		return ResponseEntity.ok().body(result);
+		if (result != null) {
+			return ResponseEntity.ok().body(result);
+		} else {
+			return ResponseEntity.notFound().build();
+		}
 	}
 
 	@PutMapping(value = "/commentaire", produces = { "application/json", "application/xml" }, consumes = {
 			"application/json", "application/xml" })
-	public ResponseEntity<Void> update(@RequestBody Commentaire p) {
-		commentaireService.update(p);
-		return ResponseEntity.noContent().build();
+	public ResponseEntity<Void> update(@RequestBody Commentaire commentaire) {
+		Commentaire existingCommentaire = commentaireService.findById(commentaire.getId());
+		if (existingCommentaire != null) {
+			commentaireService.update(commentaire);
+			return ResponseEntity.noContent().build();
+		} else {
+			return ResponseEntity.notFound().build();
+		}
 	}
 
 	@DeleteMapping(value = "/commentaire/{id}", produces = { "application/json", "application/xml" })
 	public ResponseEntity<Void> delete(@PathVariable int id) {
-		commentaireService.delete(id);
-		return ResponseEntity.noContent().build();
+		Commentaire existingCommentaire = commentaireService.findById(id);
+		if (existingCommentaire != null) {
+			commentaireService.delete(id);
+			return ResponseEntity.noContent().build();
+		} else {
+			return ResponseEntity.notFound().build();
+		}
 	}
 }
