+++ src/main/java/ma/emsi/blog/controller/ProprietaireController.java
@@ -3,6 +3,7 @@ package ma.emsi.blog.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import ma.emsi.blog.model.Proprietaire;
@@ -41,8 +43,12 @@ public class ProprietaireController {
 	@GetMapping(value = "/proprietaire/id/{id}", produces = { "application/json", "application/xml" }, consumes = {
 			"application/json", "application/xml" })
 	public ResponseEntity<Proprietaire> findProprietaire(@PathVariable int id) {
-		Proprietaire result = proprietaireService.findById(id);
-		return ResponseEntity.ok().body(result);
+		Proprietaire proprietaire = proprietaireService.findById(id);
+		if (proprietaire != null) {
+			return ResponseEntity.ok().body(proprietaire);
+		} else {
+			return ResponseEntity.notFound().build();
+		}
 	}
 
 	@PutMapping(value = "/proprietaire", produces = { "application/json", "application/xml" }, consumes = {
@@ -52,10 +58,15 @@ public class ProprietaireController {
 		return ResponseEntity.noContent().build();
 	}
 
-	@DeleteMapping(value = "/proprietaire/{id}", produces = { "application/json", "application/xml" })
+	@DeleteMapping(value = "/proprietaire/{id}")
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	public ResponseEntity<Void> delete(@PathVariable int id) {
-		proprietaireService.delete(id);
-		return ResponseEntity.noContent().build();
+		if (proprietaireService.existsById(id)) {
+			proprietaireService.delete(id);
+			return ResponseEntity.noContent().build();
+		} else {
+			return ResponseEntity.notFound().build();
+		}
 	}
 
 }
